improve undefined behavior case in bzhi emulation
Doing `~(popcnt >> 6)` gives us -2 when popcnt is 64, otherwise it gives us -1. ANDing this value with `1 << popcnt` is fine when `1 << 64` gives us `1`, which it may on a lot of machines, but C does not guarantee that to be the case. Instead, we could do `(popcnt >> 6) - 1` which gives us 0 when the popcnt is 64, and `(0 & undefined)` is 0. When the popcnt is not 64, we get `-1` from the revised expression, and ANDing with that works properly too.